display field worker location in map as well
